+++ tests/datalab/test_issue_manager.py
@@ -6,6 +6,33 @@ from cleanlab.datalab.issue_manager import IssueManager
 from cleanlab.datalab.factory import REGISTRY, register
 
 
+class TestCustomIssueManager:
+    @pytest.mark.parametrize(
+        "score",
+        [0, 0.5, 1],
+        ids=["zero", "positive_float", "one"],
+    )
+    def test_make_summary_with_score(self, custom_issue_manager, score):
+        summary = custom_issue_manager.make_summary(score=score)
+
+        expected_summary = pd.DataFrame(
+            {
+                "issue_type": [custom_issue_manager.issue_name],
+                "score": [score],
+            }
+        )
+        assert pd.testing.assert_frame_equal(summary, expected_summary) is None
+
+    @pytest.mark.parametrize(
+        "score",
+        [-0.3, 1.5, np.nan, np.inf, -np.inf],
+        ids=["negative_float", "greater_than_one", "nan", "inf", "negative_inf"],
+    )
+    def test_make_summary_invalid_score(self, custom_issue_manager, score):
+        with pytest.raises(ValueError):
+            custom_issue_manager.make_summary(score=score)
+
+
 def test_register_custom_issue_manager(monkeypatch):
     import io
     import sys
